backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.0
python-multipart==0.0.6

# Validation & Environment
pydantic==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0

# Stock Data
yfinance==0.2.33
pandas==2.1.3
numpy==1.24.3
plotly==5.18.0

# AI / ML
tensorflow==2.13.0
xgboost==2.0.3
prophet==1.1.5
scikit-learn==1.3.2

# Redis & Celery
redis==5.0.1
celery==5.3.4